Separate events traits from main
