+++ pumpkin/src/plugin/api/events/block/break.rs
@@ -1,11 +1,8 @@
+use pumpkin_event::event::{CancellableEvent, Event};
+use pumpkin_world::block::block_registry::Block;
 use std::sync::Arc;
 
-use pumpkin_world::block::block_registry::Block;
-
-use crate::{
-    entity::player::Player,
-    plugin::{CancellableEvent, Event},
-};
+use crate::entity::player::Player;
 
 use super::{BlockBreakEvent, BlockEvent, BlockExpEvent};
