+++ pumpkin/src/plugin/api/events/player/join.rs
@@ -1,11 +1,8 @@
+use pumpkin_event::event::{CancellableEvent, Event};
+use pumpkin_util::text::TextComponent;
 use std::sync::Arc;
 
-use pumpkin_util::text::TextComponent;
-
-use crate::{
-    entity::player::Player,
-    plugin::{CancellableEvent, Event},
-};
+use crate::entity::player::Player;
 
 use super::{PlayerEvent, PlayerJoinEvent};
