+++ pumpkin/src/world/mod.rs
@@ -13,7 +13,6 @@ use crate::{
             join::PlayerJoinEventImpl, leave::PlayerLeaveEventImpl, PlayerJoinEvent,
             PlayerLeaveEvent,
         },
-        CancellableEvent,
     },
     server::Server,
     PLUGIN_MANAGER,
@@ -25,6 +24,7 @@ use pumpkin_data::{
     sound::{Sound, SoundCategory},
     world::WorldEvent,
 };
+use pumpkin_event::event::CancellableEvent;
 use pumpkin_protocol::client::play::{CBlockUpdate, CDisguisedChatMessage, CRespawn, CWorldEvent};
 use pumpkin_protocol::{
     client::play::CLevelEvent,
